update: content for workflows page and some code cleanup
