+++ src/app/navbar/navbar.component.html
@@ -10,8 +10,8 @@
     <div class="ml-5 font-size-small font-roboto-bold font-white link-wrapper" [class.active]="activeRoute === 'dashboard'">
         <a [routerLink]="['/', namespaceTracker.activeNamespace, 'dashboard']">DASHBOARD</a>
     </div>
-    <div class="ml-5 font-size-small font-roboto-bold font-white link-wrapper" [class.active]="activeRoute === 'templates'">
-        <a [routerLink]="['/', namespaceTracker.activeNamespace, 'workflow-templates']">WORKFLOWS</a>
+    <div class="ml-5 font-size-small font-roboto-bold font-white link-wrapper" [class.active]="activeRoute === 'workflows'">
+        <a [routerLink]="['/', namespaceTracker.activeNamespace, 'workflows']">WORKFLOWS</a>
     </div>
     <div class="ml-5 font-size-small font-roboto-bold font-white link-wrapper" [class.active]="activeRoute === 'workspaces'">
         <a [routerLink]="['/', namespaceTracker.activeNamespace, 'workspaces']">WORKSPACES</a>
