+++ src/app/workflow/models.ts
@@ -3,9 +3,9 @@ import { ConfirmationDialogData } from "../confirmation-dialog/confirmation-dial
 
 export class WorkflowExecutionExtensions {
     public static activePhases = {
-        'Pending': true,
-        'Running': true
-    }
+        Pending: true,
+        Running: true
+    };
 
     public static isActive(workflow: WorkflowExecution): boolean {
         return this.activePhases[workflow.phase];
