+++ src/app/workflow/workflow-execute-dialog/workflow-execute-dialog.component.ts
@@ -117,6 +117,7 @@ export class WorkflowExecuteDialogComponent implements OnInit, OnDestroy {
         parameters: formattedParameters,
         labels: this.labels,
       },
+      workflowTemplate: this.selectedTemplate,
       cron: undefined,
     };
