+++ src/app/workflow/workflow.component.html
@@ -1,145 +1,33 @@
-<div class="page mb-0">
-  <div class="page-header">
-    <div class="page-links">
-      <a *ngIf="this.workflow?.workflowTemplate" [routerLink]="['/', this.namespace, 'workflow-templates', this.workflow.workflowTemplate.uid]" class="op-back-link">
-        <i class="fas fa-chevron-left"></i>
-        <span>Back to workflow template</span>
-      </a>
+<div class="page">
+    <div class="page-alerts mb-5">
+        <app-alert-panel></app-alert-panel>
     </div>
-    <div class="page-title-text">Workflow execution</div>
-    <app-alert-panel></app-alert-panel>
-  </div>
-
-  <div class="page-body">
-    <app-activity-bar *ngIf="workflow?.phase" [status]="workflow.phase" class="mt-4">
-      <div class="d-flex align-items-center w-100">
-        <app-status *ngIf="workflow" class="font-roboto-bold font-weight-bold" [phase]="workflow.phase"></app-status>
-        <span *ngIf="workflow && workflow.succeeded" class="ml-1"> in </span>
-        <app-clock *ngIf="workflow"
-                   class="ml-1"
-                   [startedAt]="startedAt"
-                   [finishedAt]="finishedAt">
-        </app-clock>
-
-        <div class="flex-grow-1">
+    <div class="page-header">
+        <div class="page-title">
+            <div class="page-title-text">Workflows</div>
+            <div class="page-action-bar d-flex">
+                <button mat-stroked-button color="primary" class="op-rounded op-button vertical-align-icon vertical-baseline-mat-icon-fa" [routerLink]="['/', this.namespace, 'workflow-templates', 'create']">
+                    <mat-icon [fontSet]="'fas'" [fontIcon]="'fa-desktop'"></mat-icon>
+                    <span class="ml-1">TEMPLATE BUILDER</span>
+                </button>
+                <a mat-flat-button color="accent" class="op-rounded op-button ml-3" (click)="executeWorkflow($event)">
+                    <mat-icon id="create-button" class="icon">add_circle</mat-icon>
+                    <span class="ml-1">Execute Workflow</span>
+                </a>
+            </div>
         </div>
-
-        <app-button
-                *ngIf="!workflow?.active && permissions.create"
-                color="primary"
-                (click)="runAgain()"
-                [loading]="cloning">
-          <span>RUN AGAIN</span>
-          <span class="loading">RUNNING AGAIN</span>
-        </app-button>
-      </div>
-
-      <div>
-        <button *ngIf="workflow?.active && permissions.delete" mat-button color="primary" class="font-size-regular" (click)="onTerminate()">Terminate</button>
-      </div>
-    </app-activity-bar>
-
-    <div class="op-info-box mt-4">
-      <div class="header">
-        <div class="page-title-text">
-          <ng-container *ngIf="workflow">{{workflow.name}}</ng-container>
-        </div>
-        <div class="actions">
-          <button *ngIf="permissions.update" mat-button color="primary" class="font-proxima-bold font-size-regular ml-2" (click)="onEdit()"><i class="fas fa-pencil-alt mr-2"></i> EDIT LABELS</button>
-          <button mat-button color="primary" class="font-proxima-bold font-size-regular ml-2" (click)="onShowTotalYaml()"><i class="fas fa-terminal mr-2"></i>VIEW YAML</button>
-        </div>
-      </div>
-      <div class="d-flex align-items-baseline font-roboto font-size-small color-medium-gray mt-3">
-        <div>
-          Started: <app-date *ngIf="workflow" [dateTime]="startedAt"></app-date>
-        </div>
-        <div *ngIf="workflow?.workflowStatus?.finishedAt" class="ml-5">
-          Ended: <app-date *ngIf="workflow" [dateTime]="finishedAt"></app-date>
-        </div>
-
-        <app-labels-viewer
-                *ngIf="labels?.length !== 0 || loadingLabels"
-                class="ml-2"
-                [labels]="labels"
-                [loading]="loadingLabels">
-        </app-labels-viewer>
-      </div>
-
-      <div class="op-h2 mt-4 mb-4">Parameters</div>
-      <ng-container *ngIf="this.parameters.length === 0">
-        <span class="font-roboto color-medium-gray">No parameters</span>
-      </ng-container>
-      <ng-container *ngFor="let parameter of this.parameters; let index = index">
-        <app-workflow-parameter
-                *ngIf="index == 0 || showAllParameters"
-                [parameter]="parameter"
-                class="d-block">
-        </app-workflow-parameter>
-      </ng-container>
-
-      <div *ngIf="this.parameters.length > 0" class="font-roboto font-size-regular underline pointer-hover mt-3" (click)="this.toggleShowParameters()">
-        <span *ngIf="showAllParameters">Hide parameters</span>
-        <span *ngIf="!showAllParameters">Show all parameters</span>
-      </div>
     </div>
 
-    <div id="info-box" class="op-info-box mt-4">
-      <div #pageContent>
-        <app-dag
-          class="mt-4 h-100"
-          [style.height]="height"
-          [style.minHeight]="'500px'"
-          (nodeClicked)="handleNodeClicked($event)"
-          [identifier]="dagIdentifier">
-        </app-dag>
-      </div>
+    <div class="page-body mt-3 workflow-list">
+        <app-workflow-executions
+                [namespace]="namespace"
+                [phase]="workflowExecutionPhase"
+                [displayedColumns]="displayedColumns"
+                [pageSize]="pageSize"
+                (workflowsChanged)="onWorkflowsChanged($event)">
+        </app-workflow-executions>
+        <app-call-to-action *ngIf="hasWorkflowExecutions !== undefined && !hasWorkflowExecutions">
+            <div class="header text-center d-block">No workflow executions found</div>
+        </app-call-to-action>
     </div>
-    <app-node-info
-       class="ml-3 op-border scroll-overflow"
-       [ngClass]="{ 'visible': showNodeInfo, 'hidden': !showNodeInfo }"
-       [namespace]="namespace"
-       [name]="uid"
-       [visible]="showNodeInfo"
-       [style.height]="nodeInfoHeight"
-       [style.minHeight]="nodeInfoHeight"
-       [style.top]="nodeInfoTop"
-       [workflow]="workflow"
-       (closeClicked)="onNodeInfoClosed()"
-       (yamlClicked)="onYamlClicked()"
-       (logsClicked)="onLogsClicked()">
-    </app-node-info>
-  </div>
-  <app-logs
-    *ngIf="showLogs && nodeInfo"
-    [namespace]="namespace"
-    [workflowName]="workflow.name"
-    [podId]="nodeInfo.id"
-    [nodeInfo]="nodeInfo"
-    (closeClicked)="onLogsClosed()">
-  </app-logs>
-  <div [ngClass]="{ 'visible': showYaml, 'hidden': !showYaml }" class="yaml-root">
-    <div class="d-flex align-items-center toolbar font-roboto">
-      <div class="title font-white">
-        <strong>YAML: </strong>
-        <span class="node-name">{{workflow?.name}}</span>
-      </div>
-      <div class="flex-grow-1"></div>
-      <button mat-icon-button class="close-button font-white" aria-label="close icon" (click)="onYamlClose()">
-        <mat-icon>close</mat-icon>
-      </button>
-    </div>
-    <ace-editor
-        #yamlEditor
-        class="w-100 h-100"
-        [mode]="'yaml'"
-        [readOnly]="true"
-        [text]="workflow?.yamlManifest"
-        [options]="{
-                    showInvisibles: false,
-                    showPrintMargin: false,
-                    scrollPastEnd: 1,
-                    showLineNumbers: true
-                  }"
-    ></ace-editor>
-  </div>
 </div>
